+++ lib/src/constants/app_strings.dart
@@ -1,6 +1,8 @@
 class AppStrings{
-  static const String notes = 'Notes';
-  static const String favourites = 'Favourites';
+  static const String allNotes = 'All Notes';
+  static const String bookmarkedNotes = 'Bookmarked Notes';
+  static const String bookmarks = 'Bookmarks';
+  static const String all = 'All';
   static const String tags = 'Tags';
   static const String noNotes = 'Oops! No notes here.';
 }
