+++ lib/src/screens/favourite_notes.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:note_it/src/constants/app_strings.dart';
 import 'package:note_it/src/widgets/no_notes_info.dart';
 
 class FavouriteNotesScreen extends StatelessWidget {
@@ -6,7 +7,7 @@ class FavouriteNotesScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Favourite Notes'),
+        title: Text(AppStrings.bookmarkedNotes),
         centerTitle: true,
         actions: <Widget>[
           IconButton(icon: Icon(Icons.search), onPressed: () {}),
